Expire access token flow handling
