+++ app/src/utils/initialManager.js
@@ -2,13 +2,17 @@ import React from 'react';
 import apiManager from './apiManager';
 import { connect } from 'react-redux';
 import { updatetWishListCount } from '../actions/app/actions';
+import { getCookie } from './utilityManager';
+import appCookie from './cookie';
 import {
   guestLoginAPI,
   storeId,
   accessToken,
   accessTokenCookie,
   wishlistDataCookie,
+  wishlistIdCookie,
   wishListCountApi,
+  logoutAPI,
 } from '../../public/constants/constants';
 import { resolveTheWishlistData } from './utilityManager';
 
@@ -36,8 +40,29 @@ export function getUpdatedWishlist(wishlist) {
     .then(response => {
       resolveTheWishlistData(response.data.data);
       const wishlistCount = response.data.data.wishlistItemArray[0].wishlistItemList.length
-      console.log('Wishlist Count --- ',wishlistCount);
+      console.log('Wishlist Count --- ', wishlistCount);
       wishlist.props.updatetWishListCount(wishlistCount);
     })
-    .catch(error => {});
+    .catch(error => { });
+}
+
+export function logoutTheUser() {
+  apiManager.post(logoutAPI).then(response => {
+    if (response.data.status === 'success') {
+      resetTheCookiesAndData();
+    }
+    // alert('Newsletter Subscription - ' + data.status);
+  }).catch(error => {
+    console.log('newsError---', error);
+  });
+}
+
+export function resetTheCookiesAndData() {
+  //Reset all the user Cookies
+  document.cookie = `${accessTokenCookie}=;path=/;expires=''`; /* accessTokenCookie + '=' + guestToken + ',' + ';path=/home'; */
+  const json_str = JSON.stringify([]);
+  document.cookie = `${wishlistDataCookie}=${json_str};path=/;expires=''`;
+  document.cookie = `${wishlistIdCookie}=;path=/;expires=''`;
+  appCookie.set('isLoggedIn', false, 365 * 24 * 60 * 60 * 1000);
+  window.location.reload(); //In case you don't reload the page, make this use as guest user.
 }
